@@ -14,9 +14,6 @@ Wheelmap::Application.configure do
                                                         :urlencode => false
                                                       }
 
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
-
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
@@ -52,5 +49,6 @@ Wheelmap::Application.configure do
 
   config.ember.variant = :development
 
+  config.eager_load = false
 end
 
